Don't install bundled libzip
Apart from header installation, which is done by
third-party/CMakeFiles.txt, installation of the bundled libzip is not
necessary, since it is statically linked. In fact, installation can
cause a previously installed version to be used in preference to the
version in the source tree, which is harmful in itself. It can also mess
up the -I flags for the rest of HHVM; see
https://github.com/facebook/hhvm/issues/3089 for details.