Don't install bundled PCRE
Don't install the bundled libpcre, since it is harmful and unnecessary
for the same reasons as installation of libzip, which I removed in pull
request #16.

Don't compile the bundled PCRE or libzip if they are not going to
be used. HPHPSetup has been included at the relevant time, so we
already have this information. If the bundled libraries are not used, we
don't want to install header files for them, since this will could cause
errors in DSO extensions if the system has a different library version.
Plus, it helps the build go faster if you don't compile unnecessary
things.