Allow building libmbfl on Windows
The first thing this does is remove the explicit call to `sh`, because all that does is break things on Windows.
The second thing we do is inline the definition of strtonum from awk (which is public domain), to allow running the script under the version of awk provided by msysgit.