Generate the proxygen headers in a windows friendly way
We can't pass environment variables before the bash script when on Windows, so pass them as arguments instead.
This also allows setting a custom path for gperf, because requiring gperf to be in your path on Windows is a bit excessive.

This depends on [proxygen#55](https://github.com/facebook/proxygen/pull/55) being merged upstream first.
This also depends on [#82](https://github.com/hhvm/hhvm-third-party/pull/82), and will conflict with it as well, due to the proximity of the changes.