Don't try to add OPENSSL_LIBRARIES to the list of link directories
This doesn't break anything when `OPENSSL_LIBRARIES` is only a single set of files that actually exist, but in multi-target configurations, such as Visual Studio, this attempts to add `debug`, `release`, and `default` as link directories, and, because they don't exist, CMake gives an error when generating the project files.
As `OPENSSL_LIBRARIES` should only ever contain files, the call isn't needed, and removing it shouldn't break anything.