Handle building Folly for MSVC
This means headers are now included in the project file, and that there's an `auto_source_group` call.
This also switches to `HHVM_REMOVE_MATCHES_FROM_LISTS`.
And lastly, it does some minor re-ordering in the list of files removed from `files` and removes `experimental/io/IOBufQueue.cpp` from that list, as it no longer exists.