'opam' CMake target to prepare for ocaml removal
Opam can be built without ocaml, there is a sandboxing mechanism
(see `make cold`). This only rely on recompiling ocaml from scratch,
no binary dependency.
In order to get rid of ocaml, we need to make hhvm depend on opam,
not on ocaml.

This patch adds a new opam target for this purpose. Once this is done,
hhvm can be updated to only depend on opam information. Once this is
done, ocaml can be stripped from third-party.